requirements.txt
alpaca-py>=0.43.0
lightgbm>=4.0.0
lxml>=5.0.0
numpy>=2.0.0
pandas>=2.0.0
plotly>=5.0.0
pyyaml>=6.0
requests>=2.28.0
scikit-learn>=1.3.0
scipy>=1.11.0
yfinance>=0.2.50
